Moving password algorithm (test) classes into better places.
