+++ graylog2-server/src/main/java/org/graylog2/security/PasswordAlgorithmFactory.java
@@ -14,9 +14,10 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.users;
+package org.graylog2.security;
 
 import org.graylog2.plugin.security.PasswordAlgorithm;
+import org.graylog2.users.DefaultPasswordAlgorithm;
 
 import javax.annotation.Nullable;
 import javax.inject.Inject;
