+++ graylog2-server/src/main/java/org/graylog2/security/hashing/BCryptPasswordAlgorithm.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.users;
+package org.graylog2.security.hashing;
 
 import com.google.common.base.Splitter;
 import org.graylog2.plugin.security.PasswordAlgorithm;
