+++ graylog2-server/src/test/java/org/graylog2/security/hashing/BCryptPasswordAlgorithmTest.java
@@ -14,9 +14,10 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.users;
+package org.graylog2.security.hashing;
 
 import org.graylog2.Configuration;
+import org.graylog2.security.hashing.BCryptPasswordAlgorithm;
 import org.junit.Before;
 import org.junit.Test;
