+++ graylog2-server/src/test/java/org/graylog2/security/hashing/SimpleHashPasswordAlgorithmTest.java
@@ -14,8 +14,9 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.users;
+package org.graylog2.security.hashing;
 
+import org.graylog2.security.hashing.SimpleHashPasswordAlgorithm;
 import org.junit.Before;
 import org.junit.Test;
