+++ graylog2-server/src/test/java/org/graylog2/users/UserServiceImplTest.java
@@ -27,6 +27,8 @@ import org.graylog2.database.MongoConnection;
 import org.graylog2.database.MongoConnectionRule;
 import org.graylog2.plugin.database.users.User;
 import org.graylog2.plugin.security.PasswordAlgorithm;
+import org.graylog2.security.PasswordAlgorithmFactory;
+import org.graylog2.security.hashing.SimpleHashPasswordAlgorithm;
 import org.joda.time.DateTimeZone;
 import org.junit.Before;
 import org.junit.ClassRule;
